requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-dotenv
apscheduler
pytest
pytest-asyncio
httpx
pydantic-settings
certifi
email-validator
transformers==4.56.1
torch
assemblyai
cloudinary>=1.36.0
python-multipart>=0.0.6
Pillow>=10.1.0
pytz